% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/once.R
\name{once}
\alias{once}
\title{Trigger an expression once a condition is met.}
\usage{
once(expr, then, envir = parent.frame(), ...)
}
\arguments{
\item{expr}{the expression to check}

\item{then}{the expression to be execuated once}

\item{envir}{the environment in which the expressions are to be evaluated.}

\item{...}{other parameters passed to \code{wait_until}.}
}
\description{
Repeatedly examine the given expression until it returns \code{TRUE}. Then trigger the evaluation
of another expression.
}
