% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retry.R
\docType{package}
\name{retry-package}
\title{retry: Retry an Expression}
\description{
The retry package provide an easy mechanism to keep evaluating an expression until either it succeeds or timeout. It is useful in situations that random failures could happen, examples would be network or file IO.
}
\author{
\strong{Maintainer}: Randy Lai \email{randy.cs.lai@gmail.com}

}
