% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retry.r
\name{retry_f}
\alias{retry_f}
\title{Retry a function}
\usage{
retry_f(f, config)
}
\arguments{
\item{f}{function to add retry behavior}

\item{config}{named list to specify retry behavior}
}
\value{
a function with retry behavior added to the function \code{f}
}
\description{
The function \code{f} is given retry behavior based on the specification in
the \code{config} configuration list.
}
\details{
The default behavior is to retry the function \code{f} forever without
waiting.
}

