% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retry.R
\name{wait_until}
\alias{wait_until}
\title{Block the current runtime until the expression returns \code{TRUE}.}
\usage{
wait_until(expr, envir = parent.frame(), timeout = Inf, interval = 0.1)
}
\arguments{
\item{expr}{the expression to check}

\item{envir}{the environment in which the expression is to be evaluated.}

\item{timeout}{raise an error if this amount of time in second has passed.}

\item{interval}{delay between retries.}
}
\description{
Block the current runtime until the expression returns \code{TRUE}.
}
\examples{

s <- Sys.time()
system.time(wait_until(Sys.time() - s > 1))

}
