% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retry.R
\docType{package}
\name{retry-package}
\title{retry: Repeated Evaluation}
\description{
Provide simple mechanism to repeatedly evaluate an expression until either it succeeds or timeout exceeded. It is useful in situations that random failures could happen.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/randy3k/retry}
}

}
\author{
\strong{Maintainer}: Randy Lai \email{randy.cs.lai@gmail.com}

}
