% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retry.R
\name{retry}
\alias{retry}
\title{Retry code block until it succeeds}
\usage{
retry(block, attempts = 10)
}
\arguments{
\item{block}{Code block to retry}

\item{attempts}{Maximum number of times to retry}
}
\value{
The result of the code block if it succeeds after the number of
    attempts, the last error it raised if it does not.
}
\description{
Retry code block until it succeeds
}
\examples{
y <- retry({
  random_num <- runif(1)
  print(random_num)
  stopifnot(random_num < .2)
  random_num
})
print(y)

z <- retry({
  random_num <- runif(1)
  print(random_num)
  stopifnot(random_num < .02)
  random_num
  }, 10
)
print(z)

}
